+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.15] — 2026-08-11
+
 ### Fixed
 
 - Ecto runner no longer emits colliding heads when a parameter is inferred as